Use "did-stop-loading" event in "continue"
The formerly used "did-finish-load" is not always emitted after an action
in case webContents.isLoading() is true. Thus the "continue" event is never
emitted. The probability of this behavior rises with rising frequency
of actions, resulting in flaky behavior and hangs depending on the
target site.

The electron documentation states that "did-finish-load" is only
dispatched when the "onload" event is dispatched and the spinner of the
tab has stopped spinning. Yet the "onload" event is not always dispatched
when webContents.isLoading() is true. Therfore the correct event is
"did-stop-loading" which is always dispatched after isLoading()
transitions to false.